Extract Japanese layout to stenoLayouts file
